Move reusable UI protocols to UILibrary
