+++ Passepartout/Library/Sources/AppUIMain/Views/Modules/Extensions/IPModule+Extensions.swift
@@ -25,9 +25,10 @@
 
 import PassepartoutKit
 import SwiftUI
+import UILibrary
 
 extension IPModule.Builder: ModuleViewProviding {
-    func moduleView(with editor: ProfileEditor, impl: ModuleImplementation?) -> some View {
+    public func moduleView(with editor: ProfileEditor, impl: ModuleImplementation?) -> some View {
         IPView(editor: editor, module: self)
     }
 }
